+++ Sources/StreamChatUI/Appearance+Formatters/MarkdownFormatter.swift
@@ -43,32 +43,137 @@ open class DefaultMarkdownFormatter: MarkdownFormatter {
     }
 
     open func format(_ string: String) -> NSAttributedString {
-        let markdownFormatter = SwiftyMarkdown(string: string)
-        modify(swiftyMarkdownFont: markdownFormatter.code, with: styles.codeFont)
-        modify(swiftyMarkdownFont: markdownFormatter.body, with: styles.bodyFont)
-        modify(swiftyMarkdownFont: markdownFormatter.link, with: styles.linkFont)
-        modify(swiftyMarkdownFont: markdownFormatter.h1, with: styles.h1Font)
-        modify(swiftyMarkdownFont: markdownFormatter.h2, with: styles.h2Font)
-        modify(swiftyMarkdownFont: markdownFormatter.h3, with: styles.h3Font)
-        modify(swiftyMarkdownFont: markdownFormatter.h4, with: styles.h4Font)
-        modify(swiftyMarkdownFont: markdownFormatter.h5, with: styles.h5Font)
-        modify(swiftyMarkdownFont: markdownFormatter.h6, with: styles.h6Font)
-        return markdownFormatter.attributedString()
+        if #available(iOS 15, *) {
+            do {
+                let attributedString = try MarkdownParser.style(
+                    markdown: string,
+                    attributes: AttributeContainer(defaultAttributes),
+                    inlinePresentationIntentAttributes: inlinePresentationIntentAttributes(for:),
+                    presentationIntentAttributes: presentationIntentAttributes(for:in:)
+                )
+                return NSAttributedString(attributedString)
+            } catch {
+                log.debug("Failed to parse string for markdown: \(error.localizedDescription)")
+            }
+        }
+        return NSAttributedString(
+            string: string,
+            attributes: defaultAttributes
+        )
     }
 
-    private func modify(swiftyMarkdownFont: FontProperties, with font: MarkdownFont) {
-        if let fontName = font.name {
-            swiftyMarkdownFont.fontName = fontName
+    // MARK: - Styling Attributes
+    
+    private var colorPalette: Appearance.ColorPalette { Appearance.default.colorPalette }
+    
+    private var defaultAttributes: [NSAttributedString.Key: Any] {
+        [
+            .font: UIFont.font(forMarkdownFont: styles.bodyFont),
+            .foregroundColor: styles.bodyFont.color ?? Appearance.default.colorPalette.text
+        ]
+    }
+    
+    @available(iOS 15, *)
+    private func inlinePresentationIntentAttributes(for inlinePresentationIntent: InlinePresentationIntent) -> AttributeContainer? {
+        switch inlinePresentationIntent {
+        case .code:
+            let attributes: [NSAttributedString.Key: Any] = [
+                // Inline currently does not have background color, although many editors prefer to do this
+                .font: UIFont.font(forMarkdownFont: styles.codeFont, monospaced: true),
+                .foregroundColor: styles.codeFont.color
+            ].compactMapValues { $0 }
+            return AttributeContainer(attributes)
+        default:
+            // emphasized etc are handled automatically by UITextView
+            return nil
         }
-        if let fontSize = font.size {
-            swiftyMarkdownFont.fontSize = fontSize
+    }
+    
+    @available(iOS 15, *)
+    private func presentationIntentAttributes(for presentationKind: PresentationIntent.Kind, in presentationIntent: PresentationIntent) -> AttributeContainer {
+        switch presentationKind {
+        case .blockQuote:
+            return AttributeContainer([
+                .foregroundColor: colorPalette.subtitleText
+            ])
+        case .codeBlock:
+            let attributes: [NSAttributedString.Key: Any] = [
+                .backgroundColor: colorPalette.background2,
+                .font: UIFont.font(forMarkdownFont: styles.codeFont, monospaced: true),
+                .foregroundColor: styles.codeFont.color
+            ].compactMapValues { $0 }
+            return AttributeContainer(attributes)
+        case .header(let level):
+            let font: UIFont
+            let foregroundColor: UIColor?
+            switch level {
+            case 1:
+                font = UIFont.font(forMarkdownFont: styles.h1Font, textStyle: .title1)
+                foregroundColor = styles.h1Font.color
+            case 2:
+                font = UIFont.font(forMarkdownFont: styles.h2Font, textStyle: .title2)
+                foregroundColor = styles.h2Font.color
+            case 3:
+                font = UIFont.font(forMarkdownFont: styles.h3Font, textStyle: .title3)
+                foregroundColor = styles.h3Font.color
+            case 4:
+                font = UIFont.font(forMarkdownFont: styles.h4Font, textStyle: .headline)
+                foregroundColor = styles.h4Font.color
+            case 5:
+                font = UIFont.font(forMarkdownFont: styles.h5Font, textStyle: .subheadline)
+                foregroundColor = styles.h5Font.color
+            default:
+                font = UIFont.font(forMarkdownFont: styles.h6Font, textStyle: .footnote)
+                foregroundColor = styles.h6Font.color ?? colorPalette.subtitleText
+            }
+            if let foregroundColor {
+                return AttributeContainer([.font: font, .foregroundColor: foregroundColor])
+            } else {
+                return AttributeContainer([.font: font])
+            }
+        case .listItem:
+            return AttributeContainer([
+                .paragraphStyle: listItemParagraphStyle(forIndentationLevel: presentationIntent.indentationLevel)
+            ])
+        default:
+            return AttributeContainer()
         }
-        if let fontColor = font.color {
-            swiftyMarkdownFont.color = fontColor
+    }
+    
+    // MARK: - Paragraph Styles
+    
+    private func listItemParagraphStyle(forIndentationLevel level: Int) -> NSParagraphStyle {
+        let style = NSMutableParagraphStyle()
+        let location = style.tabStops.first?.location ?? 28
+        style.headIndent = CGFloat(level) * location
+        return style
+    }
+}
+
+private extension UIFont {
+    static func font(
+        forMarkdownFont markdownFont: MarkdownFont,
+        textStyle: TextStyle = .body,
+        monospaced: Bool = false
+    ) -> UIFont {
+        // Default
+        var descriptor = UIFontDescriptor.preferredFontDescriptor(withTextStyle: textStyle)
+        if monospaced, let updatedDescriptor = descriptor.withDesign(.monospaced) {
+            descriptor = updatedDescriptor
         }
-        if let fontStyle = font.styling?.asSwiftyMarkdownFontStyle() {
-            swiftyMarkdownFont.fontStyle = fontStyle
+        // MarkdownFont
+        // When changing family, the descriptor should be reset
+        if let fontName = markdownFont.name {
+            descriptor = UIFontDescriptor(name: fontName, size: descriptor.pointSize)
         }
+        if let size = markdownFont.size {
+            descriptor = descriptor.withSize(size)
+        }
+        if let traits = markdownFont.styling?.symbolicTraits(), let descriptorWithTraits = descriptor.withSymbolicTraits(traits) {
+            descriptor = descriptorWithTraits
+        }
+        let font = UIFont(descriptor: descriptor, size: descriptor.pointSize)
+        return UIFontMetrics(forTextStyle: textStyle).scaledFont(for: font)
     }
 }
 
@@ -79,9 +184,6 @@ public struct MarkdownStyles {
 
     /// The font used for coding blocks in markdown text.
     public var codeFont: MarkdownFont = .init()
-
-    /// The font used for links found in markdown text.
-    public var linkFont: MarkdownFont = .init()
 
     /// The font used for H1 headers in markdown text.
     public var h1Font: MarkdownFont = .init()
@@ -118,6 +220,10 @@ public struct MarkdownFont {
         color = nil
         styling = nil
     }
+    
+    var hasFontChanges: Bool {
+        name != nil || size != nil || styling != nil
+    }
 }
 
 public enum MarkdownFontStyle: Int {
@@ -126,16 +232,16 @@ public enum MarkdownFontStyle: Int {
     case italic
     case boldItalic
 
-    func asSwiftyMarkdownFontStyle() -> FontStyle {
+    func symbolicTraits() -> UIFontDescriptor.SymbolicTraits? {
         switch self {
         case .normal:
-            return .normal
+            return nil
         case .bold:
-            return .bold
+            return .traitBold
         case .italic:
-            return .italic
+            return .traitItalic
         case .boldItalic:
-            return .boldItalic
+            return [UIFontDescriptor.SymbolicTraits.traitBold, UIFontDescriptor.SymbolicTraits.traitItalic]
         }
     }
 }